added counterparty tab for back step bugfix
