+++ src/pages/counterparties/counterparty-header.js
@@ -1,9 +1,10 @@
 import React, { useState, useEffect } from 'react';
 import { API } from 'matsumoto/src/core';
 import apiMethods from '../../core/methods';
-import { price } from 'matsumoto/src/simple';
+import { price, remapStatus } from 'matsumoto/src/simple';
 import CachedForm from 'matsumoto/src/components/form/cached-form';
 import { FieldSwitch } from 'matsumoto/src/components/form';
+import Notifications from 'matsumoto/src/stores/notifications-store';
 
 const CounterpartyHeader = ({ id }) => {
 
@@ -25,6 +26,24 @@ const CounterpartyHeader = ({ id }) => {
         });
     }, [])
 
+    const activate = () => {
+        let reason = prompt('Enter a reason');
+        API.post({
+            url: apiMethods.activateCounterparty(id),
+            body: { reason },
+            success: () => Notifications.addNotification('Counterparty activated', null, 'success')
+        });
+    }
+
+    const deactivate = () => {
+        let reason = prompt('Enter a reason');
+        API.post({
+            url: apiMethods.deactivateCounterparty(id),
+            body: { reason },
+            success: () => Notifications.addNotification('Counterparty deactivated', null, 'success')
+        });
+    }
+
     return (
         <div className="counterparty-header">
             <h1>{counterparty?.name}</h1>
@@ -33,6 +52,8 @@ const CounterpartyHeader = ({ id }) => {
                     Balance: {price(balance.currency, balance.balance)}
                 </h4>
             }
+            <h3>Status: {counterparty?.isActive ? 'Active' : 'Inactive'}</h3>
+            <h3 style={{ marginBottom: '30px' }}>State: {remapStatus(counterparty?.verificationState)}</h3>
             <CachedForm
                 render={(formik) => (
                     <div className="row">
@@ -41,6 +62,9 @@ const CounterpartyHeader = ({ id }) => {
                             value={counterparty?.isActive}/>
                     </div>
                 )}/>
+            {counterparty?.isActive ?
+                <button className="button" onClick={deactivate}>Deactivate</button> :
+                <button className="button" onClick={activate}>Activate</button>}
         </div>
     )
 }
