+++ styles/all.sass
@@ -1,5 +1,6 @@
 $background-color: #F7F8FC
 $primary-color: #FCB700
+$header-element-margin: 15px 0 15px 0
 
 .block
     h1
@@ -99,7 +100,9 @@ $primary-color: #FCB700
 .counterparty-header
     margin-top: 21px
     h4
-        margin: 15px 0 15px 0
+        margin: $header-element-margin
+    button
+        margin: $header-element-margin
 .row-group
     display: flex
     .row
